+++ HISTORY.rst
@@ -14,6 +14,9 @@ History
 - Moved all the web services URLs to a YAML file that ``ServiceURL`` class reads. It makes
   managing the new URLs easier. The file is located at ``pygeoogc/static/urls.yml``.
 - Turned off threading by default for all the services since not all web services supports it.
+- Added support for setting the request method, ``GET`` or ``POST``, for ``WFS.byfilter``,
+  which could be useful when the filter string is long.
+
 
 0.1.10 (2020-08-18)
 -------------------