+++ HISTORY.rst
@@ -10,12 +10,13 @@ History
   or more generally using any valid SQL 92 WHERE clause.
 - Added support for query with SQL WHERE clause to ``ArcGISRESTful``.
 - Changed the NLDI's URL for migrating to its new API v3.
-- Added support for CQL filter to ``WFS``, credits to `Eilio <https://github.com/emiliom>`__.
+- Added support for CQL filter to ``WFS``, credits to `Emilio <https://github.com/emiliom>`__.
 - Moved all the web services URLs to a YAML file that ``ServiceURL`` class reads. It makes
   managing the new URLs easier. The file is located at ``pygeoogc/static/urls.yml``.
 - Turned off threading by default for all the services since not all web services supports it.
 - Added support for setting the request method, ``GET`` or ``POST``, for ``WFS.byfilter``,
   which could be useful when the filter string is long.
+- Added support for asynchronous download via the function ``async_requests``.
 
 
 0.1.10 (2020-08-18)